requirements.txt
# Core dependencies
yt-dlp>=2026.2.21
moviepy>=2.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
requests>=2.32.0
audioop-lts>=0.2.0; python_version >= "3.13"
